refactor(rating): conform to types patter
